- Move documents and tags to infrastructure - Rename IDocumentService to IStorableService
