+++ src/Modules.Document/Event/AddDocumentEvent.cs
@@ -8,6 +8,8 @@
 // // --------------------------------------------------------------------------------------------------------------------
 namespace Modules.Document.Event
 {
+    using Documentania.Infrastructure.Models;
+
     using Prism.Events;
     public class AddDocumentEvent : PubSubEvent<Document>
     {
