+++ src/Modules.Document/Modules.Document.csproj
@@ -79,7 +79,6 @@
     <Reference Include="WindowsBase" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Document.cs" />
     <Compile Include="DocumentService.cs" />
     <Compile Include="DocumentModule.cs" />
     <Compile Include="Event\AddDocumentEvent.cs" />
@@ -88,7 +87,6 @@
     <Compile Include="IDocumentService.cs" />
     <Compile Include="Navigation\AllDocumentsNavigation.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Tag.cs" />
     <Compile Include="ViewModels\AllDocumentsViewModel.cs" />
     <Compile Include="ViewModels\DocumentsSubMenuViewModel.cs" />
     <Compile Include="ViewModels\NewDocumentViewModel.cs" />
