+++ src/Modules.Document/ViewModels/NewDocumentViewModel.cs
@@ -11,6 +11,8 @@ namespace Modules.Document.ViewModels
 {
     using System;
 
+    using Documentania.Infrastructure.Models;
+
     using Prism.Mvvm;
 
     public class NewDocumentViewModel : BindableBase
