+++ tst/DocumentModuleTests/NewDocumentViewModelTests.cs
@@ -3,6 +3,8 @@ using Microsoft.VisualStudio.TestTools.UnitTesting;
 
 namespace DocumentModule.Tests
 {
+    using Documentania.Infrastructure.Models;
+
     using ExAs;
 
     using Modules.Document;
